backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==5.0.1
celery==5.3.6

# ML and NLP
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.2
spacy==3.7.2
scikit-learn==1.4.0

# Resume Parsing
pyresparser==1.0.6
pdfminer.six==20221105
python-docx==1.1.0
PyPDF2==3.0.1

# Job Scraping
python-jobspy==1.1.46
beautifulsoup4==4.12.3
requests==2.31.0

# Vector Database
qdrant-client==1.7.3

# AI/LLM APIs
anthropic==0.18.1
openai==1.10.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
